refactor(camelot): seperate test tx and test setup into different files
