+++ packages/camelot/src/test-setup.ts
@@ -1,121 +1,14 @@
-import type { SwapActionParams } from '@rabbitholegg/questdk'
 import { GreaterThanOrEqual } from '@rabbitholegg/questdk'
 import { parseEther, parseUnits } from 'viem'
-import { ARBITRUM_CHAIN_ID } from './chain-ids'
-import { CAMELOT_ROUTER } from './contract-addresses'
-import { Tokens, createTestCase, type TestParams } from './utils'
-
-export const SWAP_ETH: TestParams<SwapActionParams> = {
-  transaction: {
-    chainId: 42161,
-    from: '0x2a73fa7bc21d580c01f9d983e37e15e777a73b0b',
-    hash: '0x659233e044e47a604aebd6f8dc20c385a77ed62294d8724fc98850c8ec3d7d6e',
-    input:
-      '0xb4822be3000000000000000000000000000000000000000000000000822f628b80d92acf00000000000000000000000000000000000000000000000000000000000000a00000000000000000000000002a73fa7bc21d580c01f9d983e37e15e777a73b0b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000065284bfa000000000000000000000000000000000000000000000000000000000000000200000000000000000000000082af49447d8a07e3bd95bd0d56f35241523fbab1000000000000000000000000bfbcfe8873fe28dfa25f1099282b088d52bbad9c',
-    to: '0xc873fecbd354f5a56e00e710b90ef4201db2448d',
-    value: '550000000000000',
-  },
-  params: {
-    chainId: ARBITRUM_CHAIN_ID,
-    contractAddress: CAMELOT_ROUTER,
-    tokenIn: Tokens.ETH,
-    tokenOut: '0xBfbCFe8873fE28Dfa25f1099282b088D52bbAD9C', // EQB Token
-    amountIn: GreaterThanOrEqual(parseEther('0.00055')),
-    amountOut: GreaterThanOrEqual(parseUnits('9.25', 18)),
-  },
-}
-
-export const SWAP_TOKENS: TestParams<SwapActionParams> = {
-  transaction: {
-    chainId: 42161,
-    from: '0x1a185e25636306a13d3164a511f7c610f3930caa',
-    hash: '0x44ccdf21a54f5b2ea92358fedf88da067d4b05a3f9546bfa1d3c02f006be7f13',
-    input:
-      '0xac3893ba000000000000000000000000000000000000000000000028cf73e0389658000000000000000000000000000000000000000000000000000000000000010cb66400000000000000000000000000000000000000000000000000000000000000c00000000000000000000000001a185e25636306a13d3164a511f7c610f3930caa0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000006528487400000000000000000000000000000000000000000000000000000000000000020000000000000000000000005190f06eacefa2c552dc6bd5e763b81c73293293000000000000000000000000fd086bc7cd5c481dcc9c85ebe478a1c0b69fcbb9',
-    to: '0xc873fecbd354f5a56e00e710b90ef4201db2448d',
-    value: '0',
-  },
-  params: {
-    chainId: ARBITRUM_CHAIN_ID,
-    tokenIn: '0x5190F06EaceFA2C552dc6BD5e763b81C73293293', // WOMBEX
-    tokenOut: Tokens.USDT,
-    amountIn: GreaterThanOrEqual(parseUnits('750', 18)),
-    amountOut: GreaterThanOrEqual(parseUnits('15', 6)),
-    recipient: '0x1a185e25636306A13D3164a511F7C610F3930cAa',
-  },
-}
-
-export const PARASWAP_MULTISWAP: TestParams<SwapActionParams> = {
-  transaction: {
-    chainId: 42161,
-    from: '0x09ed37c8695447ef63fc11d09012c72e69462a5f',
-    hash: '0xa233755b7f976693339b531a8356a6972529b40ddf130a58ccca4607f3a232c5',
-    input:
-      '0xa94e78ef0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000fd086bc7cd5c481dcc9c85ebe478a1c0b69fcbb900000000000000000000000000000000000000000000000000000000a93f4d8200000000000000000000000000000000000000000000000000000000a872ae3900000000000000000000000000000000000000000000000000000000a94b60a100000000000000000000000009ed37c8695447ef63fc11d09012c72e69462a5f0000000000000000000000000000000000000000000000000000000000000160000000000000000000000000353d2d14bb674892910685520ac040f560ccbc060100000000000000000000000000000000000000000000000000000000031388000000000000000000000000000000000000000000000000000000000000078000000000000000000000000000000000000000000000000000000000654aeac6780bef3a8a874278b52cfeda6514ce3000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000400000000000000000000000000000000000000000000000000000000000000320000000000000000000000000ff970a61a04b1ca14834a43f5de4533ebddb5cc80000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000020000000000000000000000000d8134acfc9c71ab51452b5ba23a31354f47390320000000000000000000000000000000000000000000000000000000000002710000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000030000000000000000000000001f721e2e82f6676fce4ea07a5958cf098d339e18000000000000000000000000000000000000000000000000000000000000271000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000c00000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000655420960000000000000000000000000000000000000000000000000000000000000028fd086bc7cd5c481dcc9c85ebe478a1c0b69fcbb9ff970a61a04b1ca14834a43f5de4533ebddb5cc8000000000000000000000000000000000000000000000000000000000000000000000000af88d065e77c8cc2239327c5edb3a432268e58310000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000020000000000000000000000000d8134acfc9c71ab51452b5ba23a31354f47390320000000000000000000000000000000000000000000000000000000000002710000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000030000000000000000000000001f721e2e82f6676fce4ea07a5958cf098d339e18000000000000000000000000000000000000000000000000000000000000271000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000c00000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000655420960000000000000000000000000000000000000000000000000000000000000028ff970a61a04b1ca14834a43f5de4533ebddb5cc8af88d065e77c8cc2239327c5edb3a432268e58310000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000',
-    to: '0xdef171fe48cf0115b1d80b88dc8eab59176fee57',
-    value: '0',
-  },
-  params: {
-    chainId: 42161,
-    contractAddress: '0xdef171fe48cf0115b1d80b88dc8eab59176fee57',
-    tokenIn: Tokens.USDT,
-    // tokenOut: Tokens.USDCE
-  },
-}
-
-export const PARASWAP_MEGASWAP: TestParams<SwapActionParams> = {
-  transaction: {
-    chainId: 42161,
-    from: '0x639d2a49bdd0cf37614f46a83e4aa49eaff1e537',
-    hash: '0xf70f6361622eafd1508791f7c5436d8c8771e41e3046804d141a10c16e32e77d',
-    input:
-      '0x46c67b6d0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000b64e280e9d1b5dbec4accedb2257a87b400db14900000000000000000000000000000000000000000000001b793377a3bf76d152000000000000000000000000000000000000000000000000000000001216c685000000000000000000000000000000000000000000000000000000001220fc5300000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000160000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000000ac000000000000000000000000000000000000000000000000000000000654b3e70df1798b42f1f4b899da8bad3471f9978000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000003800000000000000000000000000000000000000000000000000000000000000ed8000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000020000000000000000000000000fd086bc7cd5c481dcc9c85ebe478a1c0b69fcbb90000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000020000000000000000000000000d8134acfc9c71ab51452b5ba23a31354f4739032000000000000000000000000000000000000000000000000000000000000271000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000002000000000000000000000000b41dd984730daf82f5c41489e21ac79d5e3b61bc000000000000000000000000000000000000000000000000000000000000271000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000400000000000000000000000000000000000000000000000000000000000000001000000000000000000004dee4c7cc7ddbc0530cafde9cfbc55f7583fe71c717100000000000000000000000000000000000000000000000000000000000018380000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000400000000000000000000000000000000000000000000000000000000000000320000000000000000000000000eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000020000000000000000000000000d8134acfc9c71ab51452b5ba23a31354f47390320000000000000000000000000000000000000000000000000000000000002710000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000030000000000000000000000001b81d678ffb9c0263b24a97847620c99d213eb14000000000000000000000000000000000000000000000000000000000000271000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000c0000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000400000000000000000000000000000000000000000000000000000000065542490000000000000000000000000000000000000000000000000000000000000002bb64e280e9d1b5dbec4accedb2257a87b400db1490009c482af49447d8a07e3bd95bd0d56f35241523fbab1000000000000000000000000000000000000000000000000000000000000000000fd086bc7cd5c481dcc9c85ebe478a1c0b69fcbb90000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000020000000000000000000000000248009f6f78b5acd5ea2fc86d56b77a7aeee2efc000000000000000000000000000000000000000000000000000000000000271000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000003000000000000000000000000eff23b4be1091b53205e35f3afcd9c7182bf3062000000000000000000000000000000000000000000000000000000000000271000000000000000000000000000000000000000000000000000000000000000a000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000020000000000000000000000000d5b927956057075377263aab7f8afc12f85100db0000000000000000000000000000000000000000000000000000000000000000',
-    to: '0xdef171fe48cf0115b1d80b88dc8eab59176fee57',
-    value: '0',
-  },
-  params: {
-    chainId: 42161,
-    tokenIn: '0xB64E280e9D1B5DbEc4AcceDb2257A87b400DB149', // LVL Token
-    tokenOut: Tokens.USDCE,
-  },
-}
-
-export const PARASWAP_SIMPLESWAP: TestParams<SwapActionParams> = {
-  transaction: {
-    from: '0xcdd37ada79f589c15bd4f8fd2083dc88e34a2af2',
-    hash: '0xf031ddefbb6f4a441de1b127b956285871a44419948fa3b27ec3bc5392b81d5b',
-    input:
-      '0x54e3f31b0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000fd086bc7cd5c481dcc9c85ebe478a1c0b69fcbb9000000000000000000000000ff970a61a04b1ca14834a43f5de4533ebddb5cc8000000000000000000000000000000000000000000000000000000001443fd00000000000000000000000000000000000000000000000000000000001429be4c000000000000000000000000000000000000000000000000000000001443ae8700000000000000000000000000000000000000000000000000000000000001e00000000000000000000000000000000000000000000000000000000000000220000000000000000000000000000000000000000000000000000000000000038000000000000000000000000000000000000000000000000000000000000003e0000000000000000000000000cdd37ada79f589c15bd4f8fd2083dc88e34a2af2000000000000000000000000e9290c80b28db1b3d9853ab1ee60c6630b87f57e0100000000000000000000000000000000000000000000000000000000004032000000000000000000000000000000000000000000000000000000000000042000000000000000000000000000000000000000000000000000000000651e25f17fc6bc6c3c404602ad4d5307c59743a50000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000001f721e2e82f6676fce4ea07a5958cf098d339e180000000000000000000000000000000000000000000000000000000000000124c04b8d59000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000def171fe48cf0115b1d80b88dc8eab59176fee570000000000000000000000000000000000000000000000000000000065270c11000000000000000000000000000000000000000000000000000000001443fd0000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000028fd086bc7cd5c481dcc9c85ebe478a1c0b69fcbb9ff970a61a04b1ca14834a43f5de4533ebddb5cc800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000124000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000',
-    to: '0xdef171fe48cf0115b1d80b88dc8eab59176fee57',
-    value: '0',
-    chainId: 42161,
-  },
-  params: {
-    chainId: ARBITRUM_CHAIN_ID,
-    tokenIn: Tokens.USDT,
-    tokenOut: Tokens.USDCE,
-    amountIn: GreaterThanOrEqual(parseUnits('330', 6)),
-    amountOut: GreaterThanOrEqual(parseUnits('330', 6)),
-    recipient: '0xcDd37Ada79F589c15bD4f8fD2083dc88E34A2af2',
-  },
-}
-
-const PARASWAP_UNISWAP: TestParams<SwapActionParams> = {
-  transaction: {
-    chainId: 42161,
-    from: '0x0f6be61bf2cd2aafc1447b974bd9928ac6da6ea3',
-    hash: '0xc59f276cf5605428684d208c2f914d244a89e3f077f5a1196880dde69af756d1',
-    input:
-      '0xa6886da900000000000000000000000000000000000000000000000000000000000000200000000000000000000000009ed7e4b1bff939ad473da5e7a218c771d1569456000000000000000000000000fd086bc7cd5c481dcc9c85ebe478a1c0b69fcbb9000000000000000000000000e592427a0aece92de3edee1f18e0157c058615640000000000000000000000000000000000000000000000000000000023c34600000000000000000000000000000000000000000000000000000000003e747e74000000000000000000000000000000000000000000000000000000003ec4d691010000000000000000000000000000000000000000000000000000000000400000000000000000000000000000000000000000000000000000000000654b34be00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001c0000000000000000000000000000000000000000000000000000000000000022050edb4b4c79a4897af372e5f721dddf800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000002b9ed7e4b1bff939ad473da5e7a218c771d1569456000bb8fd086bc7cd5c481dcc9c85ebe478a1c0b69fcbb90000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000',
-    to: '0xdef171fe48cf0115b1d80b88dc8eab59176fee57',
-    value: '0',
-  },
-  params: {
-    chainId: 42161,
-    contractAddress: '0xdef171fe48cf0115b1d80b88dc8eab59176fee57',
-    tokenIn: '0x9ed7E4B1BFF939ad473dA5E7a218C771D1569456', // REUNI token
-  },
-}
+import { Tokens, createTestCase } from './utils'
+import {
+  SWAP_ETH,
+  SWAP_TOKENS,
+  PARASWAP_SIMPLESWAP,
+  PARASWAP_MULTISWAP,
+  PARASWAP_MEGASWAP,
+  PARASWAP_UNISWAP,
+} from './test-transactions'
 
 export const passingTestCases = [
   createTestCase(SWAP_ETH, 'when swapping ETH for tokens'),
